+++ sdk/purestorageblock/arm-purestorageblock/samples-dev/storagePoolsEnableAvsConnectionSample.ts
@@ -15,7 +15,7 @@ async function storagePoolsEnableAvsConnection(): Promise<void> {
   const subscriptionId = "BC47D6CC-AA80-4374-86F8-19D94EC70666";
   const client = new BlockClient(credential, subscriptionId);
   await client.storagePools.enableAvsConnection("rgpurestorage", "storagePoolname", {
-    clusterResourceId: "tghkgktlddwlszbeh",
+    sddcResourceId: "tghkgktlddwlszbeh",
   });
 }